refactor parse page data to common tenders processors
